Add optimizer to find the dimensions of the best rectangle after checking all rotations
Add a basic Convex Hull based approach to find the rectangle with the smallest area that
contains the vertices and extremal points of a profile. Add a parameter to allow the user
to select whether they want to optimize for the best rectangle, or whether they just want
to use the naive obvious rectangle.